requirements.txt
fdb
pyjwt
pydantic
asyncpg
fastapi
uvicorn
